backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.7.1
pydantic-settings==2.2.1
PyJWT[crypto]==2.8.0
cryptography>=42.0.0
PyMuPDF==1.27.2.2
python-multipart==0.0.9
chromadb==0.6.3
supabase==2.4.6
python-dotenv==1.0.1
httpx==0.27.0
langchain-text-splitters>=0.2.0
sentence-transformers>=3.0.0
# ── Phase 3: LLM routing ──────────────────────────────────────────────────────
# openai SDK handles both OpenRouter (free-tier) and DeepSeek BYOK —
# both are OpenAI-compatible APIs. No other LLM SDKs needed.
openai>=1.30.0
